# web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/results/gallery": {
            "get": {
                "description": "Get a paginated list of results.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Results"
                ],
                "summary": "Gallery",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The page to load.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "A comma separated list of technologies to filter by.",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "A comma separated list of technologies to filter by.",
                        "name": "technologies",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "A comma separated list of HTTP status codes to filter by.",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Order the results by perception hash.",
                        "name": "perception",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include failed screenshots in the results.",
                        "name": "failed",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Hide duplicate final URLs.",
                        "name": "hide_duplicates",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.galleryResponse"
                        }
                    }
                }
            }
        },
...
        "/results/visit": {
            "post": {
                "description": "Sets result as visited by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Results"
                ],
                "summary": "Set result as visited",
                "parameters": [
                    {
                        "description": "The result ID to set as visited",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.visitResultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.galleryContent": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "boolean"
                },
                "file_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "probed_at": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "screenshot": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "technologies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "visited": {
                    "type": "boolean"
                }
            }
        },
        "api.galleryResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.galleryContent"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "api.visitResultRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    }
}